Refactor (frontend): typed Next.js config
